club, mypage 스키마 리펙토링
